i16

+++ 2021/rust/day-05/src/lib.rs
@@ -9,8 +9,8 @@ use nom::{
 
 #[derive(Debug, Ord, PartialOrd, Eq, PartialEq)]
 pub struct Point {
-    pub x: i32,
-    pub y: i32,
+    pub x: i16,
+    pub y: i16,
 }
 
 pub fn point(input: &str) -> IResult<&str, Point> {
